Split Rx, Tx registers
