+++ upsat-comms-workspace/comms/Inc/cc_rx_init.h
@@ -1,0 +1,29 @@
+/*
+ * cc_rx_init.h
+ *
+ *  Created on: May 11, 2016
+ *      Author: manos
+ */
+
+#ifndef INC_CC_RX_INIT_H_
+#define INC_CC_RX_INIT_H_
+
+#include "cc_definitions.h"
+#include "cc112x_spi.h"
+#include <stdint.h>
+
+void rx_manualCalibration();
+
+
+void rx_registerConfig();
+
+
+
+#endif /* INC_CC_RX_INIT_H_ */
+
+
+
+
+
+
+
